+++ Sources/Taps/Harnesses/PrintHarness.swift
@@ -1,5 +1,6 @@
 import TestHarness
 
 public extension TapsHarness {
+  /// A `TapsHarness`, that prints out all `TapOutput` as TAP13.
   public static var printHarness = tapsHarnessFrom(test: .printHarness())
 }