+++ Sources/Taps/Tapes/RxGeneric.swift
@@ -2,6 +2,13 @@ import RxSwift
 import TestHarness
 
 public extension RxTaps {
+  /// A tester for common observables that will complete.
+  ///
+  /// - Parameter test: The name of the test.
+  /// - Parameter directive: The `Directive` that shall be applied. If `TestPoint`s fail, they won't break the build.
+  /// - Parameter timeout: The interval defining the maximum duration the test may need.
+  /// - Parameter scheduler: The scheduler for the test. Visit RxSwift for more info.
+  /// - Parameter observable: An `Observable` that runs your tests.
   public func test<T>(
     _ title: String? = nil,
     directive: Directive? = nil,