+++ Sources/Taps/Tester/NotEqual.swift
@@ -19,6 +19,15 @@ public extension Test {
     }
   }
 
+  /// Emits a `TestPoint`, that passes if both values are not equal.
+  ///
+  /// ```swift
+  /// tape.test("test equality", plan: 1) { t in
+  ///   t.notEqual(1, 2, "no false positives")
+  /// }
+  /// ```
+  ///
+  /// - Parameter message: The message for the test.
   public func notEqual<T: Equatable>(
     _ given: T,
     _ expected: T,
@@ -32,6 +41,21 @@ public extension Test {
     report.onNext(assertNotEqual(given, expected, message: message, source: location))
   }
 
+  /// A curried version of `Test.notEqual`. Useful for `Observable`s.
+  ///
+  /// ```swift
+  /// tape.test("just emits", timeout: 0.01) { t in
+  ///   Observable.just(3)
+  ///     .map { $0 + 1 }
+  ///     .test(
+  ///       onNext: t.notEqual(to: 3, "just emits 3"),
+  ///       onError: t.fail(with: "just won't throw'")
+  ///   )
+  /// }
+  /// ```
+  ///
+  /// - Parameter expected: The expected value.
+  /// - Parameter message: The message for the test.
   public func notEqual<T: Equatable>(
     to expected: T,
     message: String? = nil,